// web/src/types/CreateFeed.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { DisplayMode } from "./DisplayMode";

export type CreateFeed = { title: string, url: string, ttl: number, folder: string | null, filters: Array<string> | null, display_mode: DisplayMode, };